+++ db/migrate/20161111180429_remove_description_from_designs.rb
@@ -1,0 +1,5 @@
+class RemoveDescriptionFromDesigns < ActiveRecord::Migration[5.0]
+  def change
+    remove_column :designs, :description, :text
+  end
+end